Removed the BSD3 license
I have not got any written confirmation that the code to be put in
this repo can be released with BSD3 license, so I remove it and put
the licensing subjet on hold.

At the moment only my stuff (facecompressor) has a license, which
is the main reason why it is linked here as git submodule. I think
the git submodule mechanism is a good way of giving everyone
ownership of their own code and freedom to license or not as open
source.